Move to DRP pipe workflow definitions.
Add the scripts that expands relative to full paths in refcats.
This fixes the issue related to the transition to Gaia DR3
catalogs.

AP pipe seems to have added some kind of real bogus classification
task and I couldn't figure out what preSources are so I gave up.
This left the previous workflow broken. Using the DRP workflow
seems to not suffer from the same problems.